Move common code into library
