+++ src/lib.rs
@@ -1,1 +1,3 @@
+pub mod error;
 pub mod persistent_config;
+pub mod s3;
